Refactor mock net module and tests
Added flag to simulate behavior of client connection. Added mock socket class with write() method. createServer mock takes a callback function with a socket parameter. Added on event emitter function to mock module. Refactored tests and added tests to check for behavior of server when a connection is established and a server's response to a client.