Once a Connection is Established, Respond with 'Connection Established' (#4)
* Update project config file

Ran npm update to fix jest performance issues.

* Refactor server class

Added 'server' and 'socket' properties to expose the instances of net.Server & net.Socket for reference in unit tests.

* Refactor mock net module and tests

Added flag to simulate behavior of client connection. Added mock socket class with write() method. createServer mock takes a callback function with a socket parameter. Added on event emitter function to mock module. Refactored tests and added tests to check for behavior of server when a connection is established and a server's response to a client.