Add readable stream and method to close client connection
Added 'data' event and callback function to log the message sent from a client to the server, and method to close the connection when the client has written to the server. Refactored log statements and added write() method stream data prompt to format the display of the TCP data sent from the server to the client and vice versa.