Refactor mock socket class and add tests
Added custom types to utilize in mock socket class and mock instance of net.Server. Refactored the MockSocket class to include the 'on()' emitter function and 'end()' socket function which simulates the behavior resulting from the refactored source code calling these methods. Refactored tests to account for changes in source code, and added new tests to test for the resulting behavior of receiving data from a client and closing a client connection once the data has been received.