Open Input and Output Stream  (#5)
* Add readable stream and method to close client connection

Added 'data' event and callback function to log the message sent from a client to the server, and method to close the connection when the client has written to the server. Refactored log statements and added write() method stream data prompt to format the display of the TCP data sent from the server to the client and vice versa.

* Refactor mock socket class and add tests

Added custom types to utilize in mock socket class and mock instance of net.Server. Refactored the MockSocket class to include the 'on()' emitter function and 'end()' socket function which simulates the behavior resulting from the refactored source code calling these methods. Refactored tests to account for changes in source code, and added new tests to test for the resulting behavior of receiving data from a client and closing a client connection once the data has been received.

* Update README.md

* Refactor server & tests

Removed console statements and refactored formatting in source code. Tests have been refactored to use new implementation.